rename: navigator -> strategy
