+++ src/main/java/model/piece/movement/JolRouteFinder.java
@@ -8,7 +8,7 @@ import static model.position.Movement.UP;
 
 import java.util.List;
 import model.position.Movement;
-import model.navigator.LimitedBasicMoveStrategy;
+import model.strategy.LimitedBasicMoveStrategy;
 import model.piece.Castle;
 import model.position.Position;
